tests/e2e/requirements.txt
# E2E test dependencies
pytest>=9.0.2
pytest-asyncio>=0.23.5
nats-py>=2.0.0
requests>=2.31.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
obstore
